cdk_stacks/requirements.txt
aws-cdk-lib==2.133.0
constructs>=10.0.0,<11.0.0
cdklabs.generative-ai-cdk-constructs==0.1.100
